+++ routes/cards.js
@@ -1,7 +1,7 @@
 /* eslint-disable consistent-return */
 const path = require('path');
 const readFilePromise = require('fs-readfile-promise');
-const { jsonHandler, noSuchDirecrory } = require('../helpers/helpers');
+const { jsonHandler } = require('../helpers/helpers');
 
 module.exports = (req, res) => {
   const buffer = readFilePromise(path.join('data', 'cards.json'));
@@ -10,5 +10,7 @@ module.exports = (req, res) => {
     .then((result) => {
       res.status(200).send(result);
     })
-    .catch(noSuchDirecrory(res));
+    .catch(() => {
+      res.status(404).json({ message: 'Запрашиваемый файл не найден' });
+    });
 };
